+++ testdata/tree/workspace/utils/src/lib.rs
@@ -1,0 +1,8 @@
+pub fn triple(a: i32) -> i32 {
+    a * 3
+}
+
+#[test]
+fn test_triple() {
+    assert_eq!(triple(3), 9);
+}